src/third_party/blink/web_tests/display-lock/lock-before-append/acquire-clone.html

<!doctype HTML>

<style>
#container {
  contain: style layout;
  width: 150px;
  height: 150px;
  background: lightblue;
}
#child {
  width: 50px;
  height: 50px;
  background: lightgreen;
}
</style>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<script>
async_test((t) => {
  async function runTest() {
    let container = document.createElement("div");
    // Lock the container, put a child into it and put it into the document.
    await container.displayLock.acquire({ timeout: Infinity }).then(() => {
      let child = document.createElement("div");
      child.id = "child";
      container.appendChild(child);

      container.id = "container";
      document.body.appendChild(container);
    });
    // Make a copy of the container and put it in the document as well.
    let cloned = container.cloneNode(true);
    document.body.appendChild(cloned);

    t.step(() => {
      assert_true(container.displayLock.locked, "original locked");
      assert_false(cloned.displayLock.locked, "cloned copy locked");
      t.done();
    });
  }
  window.onload = runTest;
});
</script>
